refactor: messages enum들을 messages package에서 관리
