+++ src/main/java/lotto/messages/OutputMessages.java
@@ -1,4 +1,4 @@
-package lotto.configuration;
+package lotto.messages;
 
 public enum OutputMessages {
     PURCHASE_COUNT_MESSAGE("개를 구매했습니다."),
